Refactor PPShape
